+++ package.json
@@ -12,6 +12,7 @@
     "@testing-library/react": "^12.1.4",
     "@testing-library/user-event": "^13.5.0",
     "axios": "^1.3.2",
+    "date-fns": "^2.29.3",
     "framer-motion": "^9.0.2",
     "normalize.css": "^8.0.1",
     "react": "^18.1.0",
