+++ src/helpers/dateFormat.js
@@ -1,0 +1,6 @@
+import { format } from 'date-fns';
+
+export const dateFormat = date => {
+  const result = format(new Date(date), 'dd/MM/yyyy');
+  return result;
+};
